pyproject.toml
[project]
name = "simplemanus"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "openai (>=1.66.3,<2.0.0)",
    "browser-use (>=0.1.40,<0.2.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "langsmith (>=0.3.18,<0.4.0)",
    "langgraph (>=0.3.17,<0.4.0)",
    "numexpr (>=2.10.2,<3.0.0)",
    "duckduckgo-search (>=7.5.2,<8.0.0)",
    "pytest-playwright (>=0.7.0,<0.8.0)",
    "markdown (>=3.7,<4.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "pypandoc (>=1.15,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
